reorganize folders so the tests pass and game runs
