+++ Scenery/announcer_chair.py
