+++ Training/observer_abilities.py
