+++ tests/test_announcer_chair.py
@@ -1,5 +1,5 @@
 import unittest
-from source.Scenery.announcer_chair import *
+from Scenery.announcer_chair import *
 
 class Helper(object):
      def get_methods(self, obj):
