+++ tests/test_carpenter_shop.py
@@ -1,5 +1,5 @@
 import unittest
-from source.Scenery.carpenter_shop import *
+from Scenery.carpenter_shop import *
 
 class CarpenterTestCase(unittest.TestCase):
