+++ tests/test_player_chair.py
@@ -1,5 +1,5 @@
 import unittest
-from source.OnStage.player_chair import *
+from OnStage.player_chair import *
 
 class Dummy(Human):
       def choose(self, board):
