+++ tests/test_referee_chair.py
@@ -1,10 +1,10 @@
 import unittest
-from source.BackStage.referee_chair import *
-from source.BackStage.judge_pit import*
-from source.Scenery.announcer_chair import *
-from source.Scenery.carpenter_shop import *
-from source.OnStage.player_chair import *
-from source.OnStage.game_table import *
+from BackStage.referee_chair import *
+from BackStage.judge_pit import*
+from Scenery.announcer_chair import *
+from Scenery.carpenter_shop import *
+from OnStage.player_chair import *
+from OnStage.game_table import *
 
 class MuteAnnouncer(Announcer):
     def show(self, statement):
