+++ tests/test_scriptographer_desk.py
@@ -1,5 +1,5 @@
 import unittest
-from source.Scenery.scriptographer_desk import *
+from Scenery.scriptographer_desk import *
 
 class ScriptographerTestCase(unittest.TestCase):
